moved functions/constraint to different files
