+++ core/src/cpu/air/memory.rs
@@ -43,6 +43,44 @@ impl CpuChip {
         opcode_selectors.is_sb + opcode_selectors.is_sh + opcode_selectors.is_sw
     }
 
+    pub(crate) fn eval_memory_address_and_acccess<AB: SP1AirBuilder>() {
+        // For operations that require reading from memory (not registers), we need to read the
+        // value into the memory columns.
+        let memory_columns = local.opcode_specific_columns.memory();
+        builder.eval_memory_access(
+            local.shard,
+            local.clk + AB::F::from_canonical_u32(MemoryAccessPosition::Memory as u32),
+            memory_columns.addr_aligned,
+            &memory_columns.memory_access,
+            is_memory_instruction.clone(),
+        );
+
+        // Check that reduce(addr_word) == addr_aligned + addr_offset.
+        builder
+            .when(is_memory_instruction.clone())
+            .assert_eq::<AB::Expr, AB::Expr>(
+                memory_columns.addr_aligned + memory_columns.addr_offset,
+                memory_columns.addr_word.reduce::<AB>(),
+            );
+
+        // Check that each addr_word element is a byte.
+        builder.slice_range_check_u8(
+            &memory_columns.addr_word.0,
+            local.shard,
+            is_memory_instruction.clone(),
+        );
+
+        // Send to the ALU table to verify correct calculation of addr_word.
+        builder.send_alu(
+            AB::Expr::from_canonical_u32(Opcode::ADD as u32),
+            memory_columns.addr_word,
+            local.op_b_val(),
+            local.op_c_val(),
+            local.shard,
+            is_memory_instruction.clone(),
+        );
+    }
+
     /// Evaluates constraints related to loading from memory.
     pub(crate) fn eval_memory_load<AB: SP1AirBuilder>(
         &self,
